test: 우승자 테스트 수정
